started refactoring Utils.IO into io.iob, io.brat etc. (wip)
