+++ tests/io/test_iob.py
@@ -1,0 +1,4 @@
+"""Tests for the module `citation_extractor.io.iob`."""
+
+# -*- coding: utf-8 -*-
+# author: Matteo Romanello, matteo.romanello@gmail.com
